avcodec/diracdec: only signal a picture is present when there is a supported Parse Code value
Data corruption may change the Parse Code to have the 0x8 bit set but
not a value that we support decoding the picture.  This caused a crash
when we try to move the ref because we didn't decode a picture so we
didn't assign that pointer.

Also log unknown Parse Codes when they are encountered as a warning.

Cases found by afl-fuzz.